Add async versions

+++ lorawan/storage/serialization/identity-serialization.h
@@ -65,7 +65,7 @@ public:
     uint32_t offset;
     uint8_t size;
     IdentityOperationRequest();
-    IdentityOperationRequest(char tag, size_t aOffset, size_t aSize, int32_t code, uint64_t accessCode);
+    IdentityOperationRequest(char tag, uint32_t aOffset, uint8_t aSize, int32_t code, uint64_t accessCode);
     IdentityOperationRequest(const unsigned char *buf, size_t sz);
     ~IdentityOperationRequest() override = default;
     void ntoh() override;
